Moving selection into its own class
